* added overwritePosition feature (Bug #5570)
A UDM property widget is now able to overwrite an existing widget
on specified tab. The position is defined by
"univentionUDMPropertyLayoutPosition". Position counting starts
top left with "1"! If specified position is smaller than 1 or
larger than the number of widgets,, overwritePosition will be
disabled and the UDM property will be appended.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/trunk/ucs@9799 1b283449-9f2f-0410-b571-8eb228e1a901