Don't unset empty grub/groot (Bug #9985)
If grub/groot is empty, "baseConfig.get('grub/groot', '(hd0,0)')" returns the
empty string (""), if it is unset, it returns "(hd0,0)".
This triggered a failure on a KVM virtualized machine, where vda2=(hd0,1) is
the root filesystem and vda1=(hd0,0) is swap.
As part of Bug #18863 PyGrub was patched to ignore an empty "root" statement:
In this case PyGrub searches for a partition marked as bootable or uses the
first partition.
But since grub/groot was unset, an explizit "root (hd0,0)" was inserted which
confused PyGrub, since it couldn't find the kernel and InitRamFs there.


git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/trunk/ucs@19387 1b283449-9f2f-0410-b571-8eb228e1a901