* implemented caching for group member's type (user/host/group) (Bug #20330)
During open() of a group object, for each uniqueMember the object type is
determined. The result is now cached to speed up further open() requests.
The default timeout for cache results is 300 seconds and can be set by
UCR variable
directory/manager/web/modules/groups/group/caching/uniqueMember/timeout

The caching effekt is nearly unnoticable in UDM-CLI since the cli-server
forks for each request so the cache is empty on each new request.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/trunk/ucs@20830 1b283449-9f2f-0410-b571-8eb228e1a901