Bug #23092: Rework migration handling
Offline nodes should not be shown in the list of target nodes to migrate
a domain to.

Migration is the only command, which switches to another node on
success. Therefore it passes the new node_uri in dest.
On the other hand is uses 'source' instead of 'node' like any other
command, which makes switching from node_name to node_uri more
complicated.

Use 'node' instead of 'source' and simplify the switch-over to 'dest' in
uvmm_domain_overview.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-2.4/ucs@27217 1b283449-9f2f-0410-b571-8eb228e1a901