Bug #23092: Convert from names to uri and uuid
After fail-over migration the name of the domain is no longer guarenteed
to be unique: The domain is still listed on the failed old node and in
addition on the new replacement server. This is because in case of
another failure the domain would be lost.

Lookup node by URI instead of by name.
Lookup domain by UUID instead of by name.

This also saves several calls to UVMMd, which can fail if the node is
unavailable.

QA: UCS-DVS is the only known external user. Make sure this still works:
 univention-dvs-sessionbroker/umc/modules/univention/management/console/handlers/dvs_templates/__init__.py:552:
  instance_cmd = umcp.SimpleCommand('uvmm/domain/overview', options={'group': 'default', 'node': netloc, 'domain': self.domain_info.name}, switch_module_if_required=True)
 univention-dvs-sessionbroker/umc/modules/univention/management/console/handlers/dvs/__init__.py:475:
  server_cmd = umcp.SimpleCommand('uvmm/node/overview', options={'group': 'default', 'node': server}, switch_module_if_required=True)
 univention-dvs-sessionbroker/umc/modules/univention/management/console/handlers/dvs/__init__.py:476:
  instance_cmd = umcp.SimpleCommand('uvmm/domain/overview', options={'group': 'default', 'node': server, 'domain': instance}, switch_module_if_required=True)

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-2.4/ucs@27221 1b283449-9f2f-0410-b571-8eb228e1a901