Bug #23092: Fix node_name for offline nodes
Nodes, which have never been seen online by the UVMMd get a dummy record
inside of UVMMd, which is returned to the UMC module on request. Several
fields were already filled in, but .name was missing, which resulted in
an Exception during rendering of group_overview and the search dialog.

Fill in the name by extracting it from the node-URI.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-2.4/ucs@27270 1b283449-9f2f-0410-b571-8eb228e1a901