Bug #23092: handle timeout in storage queries
For every domain the actual size of the image files and the pool are
determined by an explizit query to the storage functions(). They block
when the server is not available.

Check for a valid connection before querying the server.
Timeout after 10s.
Update node.last_try to mark server as offline.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-2.4/ucs@27284 1b283449-9f2f-0410-b571-8eb228e1a901