Bug #23092: check volume for being shared
When deleting a single volume, the images was not check for being
shared. This was implemented in 2a2a3c6a22cd6bd7fb665094d00d85f09ee7ff55
when deleting the full domain using uvmm_domain_remove_images().

Move the check to its own function and use it for
uvmm_domain_remove_images() as well.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-2.4/ucs@27286 1b283449-9f2f-0410-b571-8eb228e1a901