Bug #23092: Fix interface.target_dev
For disk devices, there is a 'target_dev' and 'target_bus' attribute,
but for interfaces there is only the 'target' attribute.

In Xen this attribute uniquely identifies the interface, but is only
ever created by libvirt after a domain is defined. It is currently not
used by the UVMM UMC module.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-2.4/ucs@27321 1b283449-9f2f-0410-b571-8eb228e1a901