Bug #23812: Add HybridISO support for PyGrub2
UCS-3.0 uses grub-mkrescure, which internally uses xorriso, which
generates a so-called "hybrid ISO": The ISO images also contains a DOS
partition table, which allows the identical ISO file to be stored on an
USB stick for booting from it. This breaks PyGrub,s ince it (wrongly)
detects the DOS partition table and uses the first partition instead of
the complete ISO file.

Add a check to detech HybridISO files and fall back to unpartitioned
operation.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.0/ucs@28253 1b283449-9f2f-0410-b571-8eb228e1a901