Bug #21988: Fix misleading locale error
Getting a translation most of the time prints the very annoying and
misleading error message 'could not translate string: "foo"'.
The cause for this message is that the translation file for the
currently selected locale is not found, and such all translations will
fail, not only that specific message.

Rewrite the function to only print that message once, because it
clutters up the logfile and makes finding real errors very hard.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.0/ucs@28371 1b283449-9f2f-0410-b571-8eb228e1a901