Bug #22668: Move .info files back on re-install.
On package remove the .info files is moved to the attic by the prerm
script. On re-install of a package, the .info file is not restored,
since dpkg preserves the conffile removal unless --force-confmiss is
given.

Generate a preinst snippet to move back the old .info file from the
attic before package installation.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.0/ucs@28835 1b283449-9f2f-0410-b571-8eb228e1a901