Bug #24393: Fix Response-object reuse
When multiple Requests are send to the UMC server in short succession,
they re-use the same connection. The same goes for the Response.
When the first Requests is delayed (in this case the DC Slave was not
yet joined, so authentication was probably delayed), it will be merged
with the second request into one packet. This packet is then processed
by the UMC.protocol.client and re-split, but re-uses the same Response
object for delivery, thus overwriting the first response with the
second.

Create a new Response instance for each request.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.0/ucs@28968 1b283449-9f2f-0410-b571-8eb228e1a901