Bug #24517: Delay and really kill udm-cli
Sending a signal is asynchronous: The killing process continues without
really checking that the process died. This can lead to strange errors
where an old instance still is available for the next call to udm.

Try to kill the udm-cli-processes of the previous generation in a loop
falling back to SIGKILL.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.0/ucs@29244 1b283449-9f2f-0410-b571-8eb228e1a901