Bug #25273: Fix update script download
While /etc/apt/mirror.list was comitted, the Mirror() instance was not
updated to reflect the changed repository/mirror/version/end, which
resulted in the {pre,post}up.sh script not being copied when --updateto
is used.

Re-init the UniventionMirror() instance to get the updated value of
repository/mirror/version/end.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-2.4/ucs@29987 1b283449-9f2f-0410-b571-8eb228e1a901