Bug #27149: fix non-arch test
Since non-arch is now handled as a first-class repository format, this
test fails.
Previously the test for non-arch was outside the
try-except-ConfigurationError block, which passed the ConfigurationError
on to print_component_repositories() even thou the component was not
marked as current.
Now non-arch variant is handled inside the try-except block, which now
returns an empty list, for which an internal ConfigurationError is
raised inside print_component_repositories(). So the error message
changes.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.0/ucs@33439 1b283449-9f2f-0410-b571-8eb228e1a901