Bug #23527: 63checkucrwarning: Fix Subfile parsing
univention-quota provides subfile entries for /etc/samba/smb.conf, which
are unused as long as univention-samba* is uninstalled.

Subfiles also contains a Multifile attribute, which is wrongly detected
by this test: Since Bug #17913 Subfile entries are only effective when
(at least) one Multifile entry is registered. Otherwise it's okay that
the original Debian file does not contain the UCR-warning.

Only check File and Multifile types for the UCR-warning.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.0/ucs@33528 1b283449-9f2f-0410-b571-8eb228e1a901