Bug #27149: fix absolute path concatenation
os.path.join() on two absolute paths returns only the second path, since
Python identifies that one as absolute and thus sees no need to prefix
it with the first path. This leeds to ugly failures in mirroring, where
the path to the local repository is internally stored as an absolute
path.

Convert the local repository to also only store a relative path in
.prefix.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.0/ucs@33789 1b283449-9f2f-0410-b571-8eb228e1a901