Bug #27149: Make NonArch repository first class
The variant without the directories for different architecture is
actually documented in the SDB and probably used in a lot more cases
then previously thought.
Add to formatter for the non-architecture variante and move the handling
from get_component_repositories() to _iterate_component_repositories()
to also fix the handling of component preup.sh and postup.sh scripts in
univention-errata-component-update and in mirror.py.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.0/ucs@33795 1b283449-9f2f-0410-b571-8eb228e1a901