Bug #26257: Fix update detection when repository server is unavailable
UniventionUpdater throws an ConfigurationError when the repository
server is unreachable.

update_available() missed to translate that error to an UpdateError,
which is internally used by univention-updater. Because of that the
inner handler for --check was not catching the exception and did not
return 0 to signal "no update available", but the outer handler caught
the exception and returned 1 instead for "error", which was interpreted
as "system is not up-to-date, since an update is pending".

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.0/ucs@34057 1b283449-9f2f-0410-b571-8eb228e1a901