Bug #26560: Fix crestart handling
"/etc/init.d/bind9 crestart" was only checking for "sv status" returning
"run:", which corresponds to a running bind9. In case of errors (for
example the LDAP server is not reachable) runit runs the "finished"
script to delay the next start, which then returns "finished:".
This is also returned when the service is currently shutting down, which
can be distinguished from the error case by looking for ", want down" in
addition to the "normally down" status message.

Add a function sv_running which handles "finished:.*normally down" as a
case, where bind9 is wanted running, so crestart should restart the
service.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.0/ucs@34140 1b283449-9f2f-0410-b571-8eb228e1a901