Bug #22434: 0004: Inform about missing description
This test ist disabled, because it generates too many false positives:
E.g. ldap/base is only declared in univention-base-files, but used in
many other conffiles. Since ucslint works on a single package, it can't
detect that the description is defined in a different package (there is
no need to declare a Build-Dep on all packages defining the UCR
variables used by the current package.)

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs@34415 1b283449-9f2f-0410-b571-8eb228e1a901