Bug #28104: Allow VNC host override
Introduce UCR variable 'uvmm/umc/vnc/host' to control the format of the
VNC link. IPv[46] and IPv[46]:<regex> use IPv[46] address of the host,
which also matches the optional regular expression. An empty value uses
the hostname as used in the libvirt URI. Any other value is used
directly as a constant value, which overwrites any host specific value.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs@34686 1b283449-9f2f-0410-b571-8eb228e1a901