Bug #28417: Fix atomic locking
lockf() can only be used on files opened for writing. The notifier also
called it on the transaction file opened as read-only and ignored the
error value, thus never seeing it was doing things wrong.
With the change from svn35226 the error became visible.

Change the implementation to only do locking on write and append files.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs@35258 1b283449-9f2f-0410-b571-8eb228e1a901