Bug #26058: Handle ImportError in cache load
The split of config_registry.py into multiple sub-modules invalidates
any previous cache file, because the class name is used to instanciate
the handlers on cache load.

univention-upgrade still has an UCS-3.0 version of UCR open and
continues to use it way after the update has finished. This leads to the
following taceback, because the old implementation tries to load the new
classes, which is does not know about:
  Traceback (most recent call last):
    File "/usr/sbin/univention-upgrade", line 482, in <module>
      main()
    File "/usr/sbin/univention-upgrade", line 473, in main
      univention.config_registry.handler_set( [ '%s=no' % UCR_UPDATE_AVAILABLE] )
    File "/usr/lib/pymodules/python2.6/univention/config_registry.py", line 1086, in handler_set
    File "/usr/lib/pymodules/python2.6/univention/config_registry.py", line 699, in load
  ImportError: No module named handler

Until Bug #28481 is fixed increment the version number of the cache file
to be incompatible with any previous version. Both versions will refuse
to load the other version and will replace it with their own version.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs@35572 1b283449-9f2f-0410-b571-8eb228e1a901