Bug #23990: fix headline alignment of partition grid
The installer uses most of the time UTF-8 strings. If any (translated)
string contains characters that are represented in more than one byte,
the len() function returns the number of bytes and not number of characters.

The alignment function get_col() now converts any incoming string to
unicode before it gets aligned and converts the aligned string back to
UTF-8 before returning.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs@35583 1b283449-9f2f-0410-b571-8eb228e1a901