Bug #27730: Fix allocation error in server_select
strtok() only returns a pointer into the string, which was freed at the
end of the look, which made all pointers invalid.
If the master was added as a valid alternative to the backups, the
string was duped and not freed.

strdup() all host names and free them on changes.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs@35602 1b283449-9f2f-0410-b571-8eb228e1a901