Bug #26058: Restart interfaces on all changes
With IPv6 there is the problem that an Dual stack interface needs to be
restarted on both changes to IPv4 and IPv6. The old logic tried to be
too clever and filtered out the IPv6 change and did not restart the
interface after the change.

Instead stop and start the interface on any change. This create
warnings like "eth=eth0 not known", but that is harmless and is better
visible to the user to help diagnose network setup problems.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs@35697 1b283449-9f2f-0410-b571-8eb228e1a901