Bug #27330: Call exit_handler() on regular exit
Commit svn35535 removed the call of exit_handler(0) from main() by
accident, but calling the function even on regular exists is important,
because exit_handler() calls handlers_free_all(), which is responsible
for persisting the handler status. Otherwise the modules are
uninitialized again on the next start of the listener and take very long
to initialize, because then INIT_ONLY is not set, which causes the
listener to sync every change to the disk.
This is especially slow for all temporary lock object.

exit_handler() itself also needs to be fixed, since it's called again
as a signal handler and as the regular exit function: For signals the
program must be terminated by a signal! Otherwise the exit status is
wrong which prohibits the shell from doing the right thing.
Read <http://www.cons.org/cracauer/sigint.html> for more.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs@35714 1b283449-9f2f-0410-b571-8eb228e1a901