Bug #26476: Fix ucr unregister for Multifile
ConfigHandlers.unregister() was returning the set of all handlers, which
are affected by the remouval of the .info file.
handler_unregister() was using that set instead of a set of removed
handlers to build the set of remaining handlers.
This set gets passed to ConfigHandlers.update_divert(), which used it to
undo the diversions of Multifiles, even though they were still required.

Change ConfigHandlerDiverting.uninstall_divert() to return True or
False, indicating the diversions was actually removed.

Change ConfigHandlers.unregister() to return only a set of obsolete
handlers.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs@36556 1b283449-9f2f-0410-b571-8eb228e1a901