Bug #28672: profile mode: test for EFI or bios_grub partition in profile_complete()
profile_complete() tests for a configured BIOS boot partition if EFI has
not been detected or disabled manually by profile (use_efi=no).
If EFI has been detected, the presence of an EFI system partition is
checked. If the correct one is missing, the installer switches to interactive
mode for partitioning.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs@37038 1b283449-9f2f-0410-b571-8eb228e1a901