Bug #28672: mark free space at end of disk as "reserved"
At the end of each disk about 128MiB of free space will be marked as
"reserved" and is not available for user created partitions. It is used as
reserve for increasing the size of the GPT which will also grow the backup
GPT at the end of the disk.

"max_part_end" marks the last byte of valid user partitions.

generate_freespace has been altered to create also entries for reserved
disk space.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs@37041 1b283449-9f2f-0410-b571-8eb228e1a901