Bug #28502: interactive mode: check for BIOS/EFI partition and offer automatic fix
incomplete() checks similar to profile mode, if either a BIOS boot partition
or an EFI system partition has been configured (depending on detected EFI
support status). If the partition is missing, the user will be shown a
warning dialog. The user has the option to ignore the problem. If enough
space is present as free disk space or reserved disk space, the dialog
offers to fix the missing partition automatically.

The warning dialog is shown only once, if the user ignores the problem.
The user is always able to continue its partitioning process by pressing
F12/F6 again.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs@37045 1b283449-9f2f-0410-b571-8eb228e1a901