Bug #28552: detect partitions that occupy sectors used by GPT
In some cases prepartitioned disks with MBR contain partitions that occupy
some sectors that would be used by GPT later on. An automatic conversion
of those disks is not possible.
If one of those partitions is found, the user will be warned. The disk
will be ignored by Univention Installer until the problem has been fixed
manually.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs@37083 1b283449-9f2f-0410-b571-8eb228e1a901