Bug #29257: set bootable flag in protective MBR after all parted actions
parted up to version 2.3 does not set bootable flag in protective MBR.
Some BIOS versions are looking for that flag and skip that disk in boot
process if the flag has not been found.

The installer now sets the bootable flag to all protective partitions
in MBR if type is 0xEE and a BIOS_GRUB partition has been found on that
disk.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs@37301 1b283449-9f2f-0410-b571-8eb228e1a901