Bug #30409: Always set repository server to updates.software-univention.de
Some (broken!) routers and DNS servers always return a positive answer
to DNS queries, even when no A-record exists. This breaks the detection
of a local repository server "repository-server.$domainname".

Always use updates.software-univention.de and advise the administrator
to set up the appropriate UCR variables when a local repository is
created.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs@39321 1b283449-9f2f-0410-b571-8eb228e1a901