Bug #31733: 09/35: Workaround APT caching bug
APT only uses time stamps with second granularity, which in some cases
failes to detect an updated /var/lib/dpkg/status file. APT then still
reporst an pending update, which breaks the test.

As a work around touch the status file and re-try again, if the expected
value is not returned.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.2/ucs-3.2-0@42211 1b283449-9f2f-0410-b571-8eb228e1a901