Bug #27778: UVMM validate capacity in frontend
Previously the memory size and storage volume capacities were converted
multiple times:
  JS-FE <--string--> Python-BE <--int--> UVMMd <--int--> libvirtd
This complicates parsing and validating the values in the frontend,
since the values were stored internally as strings and needed to be
converted each time, when the value was needed.

Do the conversion only when the values are displayed.
Use MappedTextBox to do the validation on input.
Use bytes internally.
Remove the conversion code from the Python backend.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.2/ucs-3.2-0@42783 1b283449-9f2f-0410-b571-8eb228e1a901