Bug #32098: UVMM storage volume naming
Using the qcow2 format on LVM volumes is not possible, so naming the
volume with the suffix '.qcow2' is a major confusion.

Only add the format as a suffix when creating storage volumes in file
based storage pools.
If there is only one valid format, automatically select it by default.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.2/ucs-3.2-0@42857 1b283449-9f2f-0410-b571-8eb228e1a901