Bug #28608: DB fix auto drop unused bibliography
bibliography.collection is included using document(), which changes the
context node, so the lookup using key('') no longer references the
original document and does not find any references using this entry.

Apply the rule only when the document includes it's own bibliography
section and does not use the include mechanism.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs-3.1-2@42918 1b283449-9f2f-0410-b571-8eb228e1a901