Bug #30962: DocBook: hide URL
By default the URL text is appended after the link in PDFs, since in a
printed document the URL is not clickable and visible otherwise:
  <ulink linkend="URL">Text</ulink>
is rendered as
  Text [URL]
This clutters the change log, since there for every bug reference a long
URL is printed.

Disable printing the URL for now.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.2/ucs-3.2-0@43329 1b283449-9f2f-0410-b571-8eb228e1a901