Bug #29337: ldiff: Handle ssh exiting early
We need to wait until data is available. There are two cases:
1. The normal case is for the remote "slapcat" to start sending data.
Later on it terminates and returns 0 for success.

2. In the error case ssh aborts and closes its STDOUT. The reader-end of
the pipe then becomes readable to signal EOF, which terminates the
select(). The process returns an error.

Basically we need to get the status of the child process, when the pipe
becomes readable: If the remote shell died by signal or error, raise an
exception to abort ldiff.
If the process is still running or terminated with success, continue
with ldiff.

But there's a race condition: The pipe is closed a little bit before the
process dies, so the pipe becomes readable while the process is still
running. Therefore we delay reading the process status by half a second,
which is hopefully enough.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.2/ucs-3.2-0@43373 1b283449-9f2f-0410-b571-8eb228e1a901