Bug #32297: Comment out 'loopback' interface from moved template
The template for /etc/network/interfaces was converted from a
single-file template to a multi-file template for UCS-3.1 and the
configuration for the loopback interface was separated into a new heder
file.

During the upgrade a modified template is moved to the new location.
Since the old template usually contains its own declaration for the lo
interface, there are now two declarations which ifup / ifdown refuses to
process.

Remove the loopback interface from the moved template.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs-3.1-1@43423 1b283449-9f2f-0410-b571-8eb228e1a901