Bug #31844: Prevent update from restarting network
"univention-config-registry register" is called on updates of
univention-base-files to register new and modified templates. This also
triggers calling the interfaces.py module, which restarts the network.
This is undesirable when using bonding or other advanced network
configurations, which can take a long time to recover.

Save the previous value of interfaces/restart/auto and set it to false
in univention-base-files.preinst, and restore the value on success in
postinst and in failures in postrm.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.2/ucs-3.2-0@43447 1b283449-9f2f-0410-b571-8eb228e1a901