Bug #29456: Keep Mercurial metadata
Switch Debian source format to "(3.0) native".

Add the Mercurial (hg) meta-data to our Subversion repository and only
strip it during package build. Otherwise the version history is lost,
which prevents looking for changed.

Also make sure to not include the privat keys, key secrets and
certificates in the generated file.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.2/ucs-3.2-0@44584 1b283449-9f2f-0410-b571-8eb228e1a901