Bug #28670: USS: Handle primary interface only change
When only the primary interface is changed, the old interface
configuration is not copied into the profile, which caused all
interfaces to be removed.

Build temporary new configuration by copying the current settings and
applying the changes.

Add a test script to easily test several profiles.

Known bug: The PXE configuration seems not be to updated when only the
primary interface changes.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.2/ucs-3.2-0@45831 1b283449-9f2f-0410-b571-8eb228e1a901