Bug #33258: USS: Fix duplicate network error
The problem is that ChangeSet.(old|new)_ipv[46]s() returns a *set* of
IPNetwork() objects. If two interfaces are configured into the same
sub-network, the second one is eliminated as a duplicate of the first
one (because the network-address of x.x.x.1/24 and x.x.x.2/24 are both
x.x.x.0/24 and thus the same).

AddressChange() only used that filtered view to detect, if there was
change in the number of configured addresses and skipped all changes:

> # grep "No address change" /var/log/univention/setup.log
> WARNING:uss.network.plug:Phase skipped: No address change

Change (old|new)_ipv[46]s() to return a list of all IP-Addresses instead
of a set and move the duplicate-network-filtering down into the
individual checks.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.2/ucs-3.2-0@47175 1b283449-9f2f-0410-b571-8eb228e1a901