Bug #34241: YAML UCR: Fix atomic write to preserve old file permission
In UCS-3.2-0 UCR overwrites any existing file and such keeps the file
permission and ownership.
This was changed for UCS-3.2-1 to do a atomic commit, which writes to a
new file first and renames that afterwards.

While the file permission was kept, the file ownership was not kept.
This only shows when no explicit owner / group is specified, but the
ownership is changed through an explicit call to "chown" by hand.

Add copy to copy the ownership information from the old file before
applying the explicitly specified changes.

Fix ucs-test to explicitly test using "chown".

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.2/ucs-3.2-1@48423 1b283449-9f2f-0410-b571-8eb228e1a901