Bug #34896 errata: reformat maximum column width
Limit to 80 columns since the text is used in a HTML-<pre>-environment.

Fix some spelling mistakes:
- JavaScript
- FreeRADIUS
- independ[a->e]nt
- applicat[ed->ion]

Add missing dot at end of sentence.

The 'note' is rendered between the description and the list of bug/CVE
Ids. As such move the line between 'desc' and 'bug'/'cve' to reflect
this, as it's otherwise unclear how the note should be phrased.
Remove all other empty notes.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.2/ucs-3.2-1@50500 1b283449-9f2f-0410-b571-8eb228e1a901