Bug #34971 uldap: Delete old RDN on rename
Since the initial CVS import uldap.rename() does not remove the old RDN
when an object is renamed. As the newer Listener makes this visible to
the S4 connector, it fails to rename some object, where the LDAP
attribute is declared as single-valued but gets multi-values through the
rename.

As most other places (replication.py) already use delold=1, do that as
well within uldap.py.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.2/ucs-3.2-2@50731 1b283449-9f2f-0410-b571-8eb228e1a901