Bug #35261 Listener: Retry update after delete by UUID [YAML]
In case a handler fails on delete, the cache entry was not deleted.

This may later cause the listener to handle the re-addition at the same
location as a delete, as the old entryUUID is still found at the cache
location and then only change_delete_dn() gets called.

Re-try the ldap search with the DN in case the first search by UUID
returned nothing.

Always delete the cache entry, as the delete handler will never get
called again anyway.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.2/ucs-3.2-2@51538 1b283449-9f2f-0410-b571-8eb228e1a901