Bug #36934 SSL: Timeout ntpdate command after 15s
ntpdate might get stuck.
Forcefully terminate the command after 15 seconds and continue without
updating the system time. This can lead to subsequent errors, but it is
still better then a stuck system.

As rdate is only used in the postinst, move the dependency from
Pre-Depends to Depends.

Also use the ntpdate-debian wrapper to use the NTP servers from the
Debian pool, which is betten then using a hard-coded server.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.0/ucs-4.0-0@56063 1b283449-9f2f-0410-b571-8eb228e1a901