Bug #32979: Do not reraise exceptions
Many exceptions were reraised so that the original traceback information
was lost when receiving e.g. external feedback with traceback reports.
Another more heavy thing was that the LDAP_Connection decorator masked
tracebacks by wrapping it into a LDAP_ConnectionError() exception.
Especially for methods which are not direct UMCP calls this removed the
necessary traceback information.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.0/ucs-4.0-0@56420 1b283449-9f2f-0410-b571-8eb228e1a901