Bug #37164 DVD: Re-add netcat and bozohttpd for errata4.0-0 DVD
The errata DVD is missing the following packages:
 netcat
 bozohttpd

This is caused by an undetermined behavior of "dselect install", which is used
to create the list of packages for the DVD: The initial status file is
generated by selecting all packages with "Priority:
(required|important|standard)".
As both "netcat-openbsd" and "netcat-traditional" are "important", they are
both chosen and are preferred to "netcat", which is only of priority "extra"
(and thus not chosen by default by the above mention algorithm):

$ LANG=C aptitude search netcat -F '%P %p'
extra     netcat
important netcat-openbsd
important netcat-traditional

Thus the package "netcat" is not put on the DVD by default.

Something similar probably happens to "bozohttpd".

For now add both packages to the task-list.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.0/ucs-4.0-0@56512 1b283449-9f2f-0410-b571-8eb228e1a901