Bug #37215 KVM: Load kernel module 'bridge'
Load kernel module 'bridge' before invoking 'ucs-kvm-setup-bridge', as
module auto loading does not work when
univention-virtual-machine-manager-node-kvm is installed in the chroot
environment during univention-system-setup.

Without the module being loaded only the address on ethX is cleared
without being transfered to brY. This results in an un-accessible
system, which stalls when the join scripts are called.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.0/ucs-4.0-0@56544 1b283449-9f2f-0410-b571-8eb228e1a901