Bug #34761 Repl: Prevent recursive listener restart YAML
If the Listener detects a schema change, it restarts the LDAP server. If
then a failed.ldif file exists, u-d-replication-resync is called to
re-apply the pending changes. It tries to stop the listener, which
fails because the listener is already executing one handler, for which
all signals are disabled. The script then gets stuck in an endless loop
waiting for the listener to shutdown, which will never happen.

1. Abort early if the failed.ldif file is
   a) not a file
   b) not readable
   c) still opened by replication.py
   d) empty file is just removed
2. Skip shutdown when called from the listener, e.g. the listener
   process is a parent process.
3. Fix quoting problems in getdns() and for ldapmodify.
4. Fix output messages
   a) fix spelling
   b) fix wrong output file name
   c) fix script name in header

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.0/ucs-4.0-0@56681 1b283449-9f2f-0410-b571-8eb228e1a901