Bug #36233 base: ignore clock skew
The installer tries to update the current time from some external time
server. During the join the time is fetched again from the DC Master,
which might be out-of-sync: If the master is behind the newly installed
system, the time goes backwards, which will break on the next file
system check:

> Checking root file system
> [....] Checking root file system...fsck from util-linux 2.20.1
> /dev/mapper/vg_ucs-root: Superblock last mount time (Mon Oct 20 11:12:43 2014,
>         now = Fri Oct  3 13:34:53 2014) is in the future.

Provide a configuration file /etc/e2fsck.conf to disable that check in e2fsck.
As a system might be joined multiple times, ignore this always and not
only while running univention-system-setup-boot.

Supports arbitrary UCR variables in the form 'e2fsck/SECTION/TAG=VALUE'.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.0/ucs-4.0-0@57212 1b283449-9f2f-0410-b571-8eb228e1a901