Bug #37509 test: 01/96admins: Skip sshd on non-DCs
Skip test for /etc/security/access-sshd.conf on member server, as the
UCRVs auth/sshd/ are not set for them.

With UCS-3.2 they were, because an appliance is setup as a DC Master
first with UCS-3, while with UCS-4 it stays without a role until fully
provisioned.

Also print all failing files instead of aborting on the first failure.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.0/ucs-4.0-0@57228 1b283449-9f2f-0410-b571-8eb228e1a901