Bug #36044 Updater: Set timeout of 600s for HTTP connections YAML
socket._GLOBAL_DEFAULT_TIMEOUT is used by default, which is a singleton
instance to implement an infinite time-out by default. This leads to the
situation, where the updater stalls infinitely when the download server
updates.software-univention.de for some yet unknown reason single-sided
terminates the connection.

Add a default-timeout of 600s=10m.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.0/ucs-4.0-1@58466 1b283449-9f2f-0410-b571-8eb228e1a901