Bug #37631 UCR: Fix univention-ldapsearch race condition YAML
If "tee" is delayed, the file is still empty and does not contain the
expected error message.
This can be forced by using the following replacement for tee:

> # cat /usr/local/bin/tee
> #!/bin/bash
> exec > >(exec /usr/bin/tee "$@")
> while read line
> do
>         sleep .1
>         echo "$line"
> done

Run a new tee process for each loop and make sure the previous instance
terminated by moving everything into a single shell pipe.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.0/ucs-4.0-1@58482 1b283449-9f2f-0410-b571-8eb228e1a901