Bug #37971 PAM: Delay auth/sshd/ restrictions until role is known YAML
Move setting the auth/sshd/ restrictions into the join script, as for
appliances the server/role isn't final when univention-pam is first
installed.

For new installations nothing changes as 11univention-pam.inst is
invoked from the .postinst.

The join script version is not incremented as we don't want to apply
those settings to existing systems.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.0/ucs-4.0-1@59502 1b283449-9f2f-0410-b571-8eb228e1a901