Bug #38154: added UCR variable system/setup/boot/force/fqdn for systems with static hostname e.g. docker instances
This commit adds the UCR variable system/setup/boot/force/fqdn. If set, the
specified value is shown/used as FQDN in Univention System Setup. The hostname
widget is disabled so the user is not able to change the value.

This feature is required e.g. for docker instances that do not allow the
modification of the hostname.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.0/ucs-4.0-1@59527 1b283449-9f2f-0410-b571-8eb228e1a901